Users can now add new lists
Users can add new lists with the 'N' key. The list created is added
after the one the user is currently on, and then current_list_index is
updated to point to the newly created list. Maybe the ESC key should
cancel creating a new list? Need to be able to delete lists now.