Users can now create .tasks.md files through the program
If the program fails to deserialise the data it will prompt the user to
provide a name for the first list. The function is a bit poorly named
but it gets the point across. If user's dont want to create a file, they
can Ctrl + C out of the program. I think this works well for now at the
very least.