Changed prompt for creating a .tasks.md file
Updated the prompt to be a simple y/n keypress, in the interest of
speed. Since users can easily change the title of the list, having a
default name isn't so bad. I haven't extensively tested it but I think it
works just fine. It does use some crossterm libraries, so I think I
might have to move the code around in the interest of separating
concerns. It does cleanup after itself though?