Users can now use arrow keys while typing as they would expect
Created the `typing_line()` function that contains and standardises
getting typing input from the user for creating and renaming/rewording
tasks and lists, and replaced the existing logic in each of those
methods. This new function also allows the use of arrow keys, allowing
the user to change where they are typing in the line. Updated version
number.