Clipboard can now hold multiple tasks
The clipboard is now a vector of tasks, so that it is capable of loading
multiple tasks into it, and paste all of them in order later. This
allows the dA command to yank all the deleted tasks. Also added the yA
command which yanks all the tasks in the current list. The clipboard is
cleared before it is written too, regardless of whether it contains one,
multiple, or no tasks.