Implemented fast list jumping
Users can now press the number key of the list they'd like to jump to,
similar to harpoon marks, so that they can quickly navigate to specific
lists. Added this feature as I found that with this project, my first
list is features to add, my second list is potential features, and my
third list is bugs, so it makes sense to be able to instantly get to the
list I'm thinking.