Fixed the issue where the cursor was placed wrong for text prompts
Previously if the task being edited was completed, the cursor would
appear way after the end of the line, yet deleting would be the same.
My theory is that escape characters are counted, so to prevent this I
did the ugly inelegant way of the function taking a parameter that
determines the length of the prefix, which is then used to place the
cursor in the correct position.